+++ bricks/illustrations/src/bootstrap.ts
@@ -1,3 +1,3 @@
 import "./data-providers/get-illustration.js";
 import "./data-providers/get-illustrations-by-category.js";
-import "./rocket-transition/index.js";
+import "./loading-transition/index.js";
