+++ bricks/illustrations/src/loading-transition/images/sprite@2x.png
